script: retire only items not part of packages

+++ script/leihs/retire_items.rb
@@ -12,11 +12,15 @@ csv_parser.for_each_row do |row|
   log("retiring #{row['Inventarcode']}", :info, true)
   item = Item.find_by_inventory_code(row['Inventarcode'])
   if item
-    item.reservations.where.not(status: 'closed').destroy_all
-    item.retired_reason = 'muss nicht mehr inventarisiert werden; ausgemustert durch skript'
-    item.retired = Date.today
-    item.save!
-    csv_parser.row_success!
+    unless item.parent_id
+      item.reservations.where.not(status: 'closed').destroy_all
+      item.retired_reason = 'muss nicht mehr inventarisiert werden; ausgemustert durch skript'
+      item.retired = Date.today
+      item.save!
+      csv_parser.row_success!
+    else
+      log("skipping #{row['Inventarcode']}, because it's part of a package", :info, true)
+    end
   else
     log("could not find #{row['Inventarcode']}", :info, true)
   end
